+++ src/sql/IChanceDAO.java
@@ -1,5 +1,11 @@
 package sql;
 
-public class IChanceDAO {
+import java.sql.SQLException;
+
+public interface IChanceDAO {
+	
+	public void getChanceCard(int cardID) throws SQLException;
+	
+	void updateChanceCard(int cardID, String cardText, int cardValue) throws SQLException;
 
 }
