+++ src/sql/IFieldDAO.java
@@ -4,13 +4,11 @@ import java.sql.SQLException;
 
 public interface IFieldDAO {
 
-	void removePlayer(int playerID) throws SQLException;
-	
-	void updatePlayer(int playerID) throws SQLException;
+	void getOwnable(int playerID) throws SQLException;
 	
 	void updateOwnable(int playerID) throws SQLException;
 	
-	void updateChanceCard(int cardID, String cardText, int cardValue) throws SQLException;
+	
 	
 	
 }
